feat: #CRRE-606 make non historical order archived during bookseller transition

+++ src/main/java/fr/openent/crre/service/impl/DefaultBasketOrderItemService.java
@@ -69,6 +69,10 @@ public class DefaultBasketOrderItemService implements BasketOrderItemService {
 
     @Override
     public Future<List<BasketOrderItem>> delete(List<Integer> basketIds) {
+        if (basketIds.isEmpty()) {
+            return Future.succeededFuture(new ArrayList<>());
+        }
+
         Promise<List<BasketOrderItem>> promise = Promise.promise();
 
         String query = "DELETE FROM " + Crre.crreSchema + ".basket_order_item WHERE id IN " + Sql.listPrepared(basketIds) + " RETURNING *";
